Genrator bug fixed - significant activity decreaseing
git-svn-id: https://svn.callweaver.org/openpbx/trunk@1518 e56c2a57-502e-0410-a6e2-86b2f1ef08c6

+++ corelib/generator.c
@@ -193,7 +193,7 @@ static void *opbx_generator_thread(void *data)
 	void (*cur_gen_free)(struct opbx_channel *chan, void *cur_gen_data);
 	struct timeval tv;
 	struct timespec ts;
-	int sleep_interval_ns;
+	long sleep_interval_ns;
 	int res;
